Insecure option
It's hard to segment vlan in a bridged interface since it's a software layer 2 switch. Even if it works in some configurations the risk of causing network issues such as loops is too big. We do not recommend this solution without proper configuration for each environment. A simple guide will not cover all scenarios